separated into playerview and spectatorview
