+++ src/components/Common/index.ts
@@ -1,10 +1,12 @@
 export { default as Avatar } from './Avatar/Avatar';
+export { default as Box } from './Box/Box';
 export { default as Button } from './Button/Button';
 export { default as Divider } from './Divider/Divider';
 export { default as Layout } from './Layout/Layout';
 export { default as Loading } from './Loading/Loading';
 export { default as LoadingFull } from './LoadingFull/LoadingFull';
 export { default as Logo } from './Logo/Logo';
+export { default as Modal } from './Modal/Modal';
 export { default as ProfileSelector } from './ProfileSelector/ProfileSelector';
 export { default as SectionTitle } from './SectionTitle/SectionTitle';
 export { default as Title } from './Title/Title';
