+++ src/components/Lobby/GameView/SpectatorView/SpectatorView.tsx
@@ -1,0 +1,5 @@
+import React from 'react'
+
+export default function SpectatorView(): JSX.Element {
+	return <p>Hey</p>
+}
