+++ src/components/Lobby/UserEtiquette/UserEtiquetteStyleBuilder.ts
@@ -1,4 +1,4 @@
-import StyleBuilder from '../../../../../server/classes/StyleBuilder';
+import StyleBuilder from '../../../../server/classes/StyleBuilder';
 
 export default class UserEtiquetteStyleBuilder extends StyleBuilder {
